+++ src/main/gui/vector/AboutBorder.hpp
@@ -2,9 +2,6 @@
 
 #include <juce_gui_basics/juce_gui_basics.h>
 
-#include "TextWithLinks.hpp"
-#include "VmpcJuceResourceUtil.hpp"
-
 namespace vmpc_juce::gui::vector {
 
     class AboutBorder : public juce::Component {
@@ -19,14 +16,14 @@ namespace vmpc_juce::gui::vector {
                 const auto lineThickness = scale;
                 const auto radius = scale * 3;
                 const auto margin = scale * marginAtScale1;
-                const auto rect = getLocalBounds().toFloat().reduced(lineThickness * 0.5).withTrimmedLeft(radius).withTrimmedRight(radius);
-                const auto top = rect.withTrimmedBottom(getHeight() - margin);
-                const auto bottom = rect.withTrimmedTop(getHeight() - margin);
+                const auto rect = getLocalBounds().toFloat().reduced(lineThickness * 0.5f).withTrimmedLeft(radius).withTrimmedRight(radius);
+                const auto top = rect.withTrimmedBottom(static_cast<float>(getHeight()) - margin);
+                const auto bottom = rect.withTrimmedTop(static_cast<float>(getHeight()) - margin);
 
                 g.setColour(juce::Colours::white);
                 g.fillRect(top); g.fillRect(bottom);
 
-                const auto rect2 = getLocalBounds().toFloat().reduced(lineThickness * 0.5);
+                const auto rect2 = getLocalBounds().toFloat().reduced(lineThickness * 0.5f);
 
                 g.setColour(juce::Colours::black);
                 g.drawRoundedRectangle(rect2, radius, lineThickness);
